// web-app/client/__generated__/globalTypes.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

//==============================================================
// START Enums and Input Objects
//==============================================================

export interface FDTaskProps {
  algorithmName: string;
  errorThreshold: number;
  maxLHS?: number | null;
  threadsCount: number;
}

export interface FileProps {
  delimiter: string;
  hasHeader: boolean;
}

//==============================================================
// END Enums and Input Objects
//==============================================================